feat moved score to own component
