+++ components/AnimeReview/index.js
@@ -1,6 +1,7 @@
 import EditIcon from "components/Icons/EditIcon"
 import Modal from "components/Modal"
 import ReviewForm from "components/ReviewForm"
+import ScoreStamp from "components/ScoreStamp"
 import { useSession } from "next-auth/react"
 import { useState } from "react"
 import styles from "./styles.module.scss"
@@ -13,11 +14,6 @@ export default function AnimeReview({
 
   const { data: session } = useSession()
 
-  const scoreStampStyles = (score) => {
-    if (score < 4) return styles.scoreStamp__red
-    if (score < 7) return styles.scoreStamp__yellow
-    return styles.scoreStamp
-  }
   const stateStampStyles = (state) => {
     if (state === "viendo") return styles.stateStamp_viendo
     if (state === "dropeada") return styles.stateStamp_dropeada
@@ -28,9 +24,7 @@ export default function AnimeReview({
       <section className={styles.card}>
         <div className={styles.score}>
           <h2 className={styles.subTitle}>score</h2>
-          <div className={scoreStampStyles(score)}>
-            <span className={styles.scoreValue}>{score}</span>
-          </div>
+          <ScoreStamp score={score} />
         </div>
         <div className={stateStampStyles(state)}>{state}</div>
         <div className={styles.watched}>
