+++ components/ScoreStamp/styles.module.scss
@@ -1,0 +1,34 @@
+@import "styles/theme.scss";
+
+.scoreStamp {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 100px;
+  width: 100px;
+  margin: 0.5rem 0;
+  border-radius: 50%;
+  border: 15px solid green;
+  background-color: limegreen;
+  box-shadow: 4px 8px 16px #0000001a;
+  color: green;
+}
+.scoreStamp__red {
+  @extend .scoreStamp;
+  border: 15px solid #800000;
+  background-color: hsl(0, 61%, 50%);
+  color: #800000;
+}
+
+.scoreStamp__yellow {
+  @extend .scoreStamp;
+  border: 15px solid hsl(60, 80%, 45%);
+  background-color: hsl(60, 60%, 70%);
+  color: hsl(60, 100%, 30%);
+}
+
+.scoreValue {
+  font-size: 35px;
+  font-weight: 800;
+  user-select: none;
+}
